# perf: optimize sample_floyd by unsafe APIs