- [x] Added a `CHANGELOG.md` entry

# Summary

This PR uses unsafe APIs to boost performance of `sample_floyd`. The optimization is totally safe because the index is bounded by the length of the vec.

# Motivation

Rust's bounds checking are sometimes unnecessary. Removing bounds checking by unsafe APIs can boost its performance.This optimization makes related functions more faster with safety ensured.

# Details

The benchmark results from my environment is listed as below. 

```
seq_slice_choose_multiple_1_of_1000
                        time:   [17.377 ns 17.480 ns 17.597 ns]
                        change: [-5.6211% -4.8379% -4.0867%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

seq_slice_choose_multiple_10_of_100
                        time:   [53.166 ns 53.654 ns 54.192 ns]
                        change: [-7.2861% -6.4089% -5.4998%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```