@dhardy 2025-04-05
Thanks for the PR.

My main concern here is simply: should we be adding more `unsafe` code for ~5% perf gains?

CC @RalfJung