@Unparalleled-Calvin 2025-04-07
Thanks for considering!
The `unsafe` block is minimized with clear `// safety` comments and encapsulated by the safe `sample_floyd` function. And tests/MIRI can confirm no UB. So I think it is ok to use unsafe code and gain more performance.