@RalfJung 2025-04-07
> Thanks for the PR.
> 
> My main concern here is simply: should we be adding more `unsafe` code for ~5% perf gains?
> 
> CC @RalfJung

Not sure what exactly you want my input on here. :)  Happy to consul on whether some use of unsafe is sound or not, but that doesn't seem to be the question here?  As to whether you think the bit of unsafe is worth the perf gain -- that's a maintainer decision. There's absolutely cases where the perf gain is important enough to justify a bit of unsafe and there are other cases where it's not worth it. I don't have to maintain this code going forward so I can't make this decision for you. :)