@Unparalleled-Calvin 2025-04-12
Thank you for your review! Here are the benchmark results of using the unsafe functions.

Only use `*indices.get_unchecked_mut(pos) = j;`  [Compared to the original version]
```
seq_slice_choose_multiple_1_of_1000
                        time:   [19.089 ns 19.212 ns 19.346 ns]
                        change: [-0.8332% +0.7077% +2.5236%] (p = 0.46 > 0.05)
                        No change in performance detected.

seq_slice_choose_multiple_10_of_100
                        time:   [58.523 ns 58.777 ns 59.060 ns]
                        change: [-6.5131% -5.2792% -4.0724%] (p = 0.00 < 0.05)
                        Performance has improved.
```
Additionally use `ptr::write` instead of `push` [compared to above]
```
seq_slice_choose_multiple_1_of_1000
                        time:   [18.896 ns 19.158 ns 19.446 ns]
                        change: [-5.8287% -3.9616% -2.4368%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)

seq_slice_choose_multiple_10_of_100
                        time:   [56.478 ns 56.843 ns 57.254 ns]
                        change: [-4.5315% -3.8087% -3.0675%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
```